refactor(WebSocketConnectionManager): 패키지 변경
